+++ rust/src/lib.rs
@@ -104,6 +104,7 @@ pub mod status;
 pub mod sys;
 pub mod sysrun;
 pub mod system;
+pub mod sysuplink;
 pub mod tools;
 pub mod trust;
 pub mod watch;